added randomCard, removed card component and added styling to card components
