+++ components/card.module.scss
@@ -6,9 +6,13 @@
   border-radius: 1rem;
   justify-content: space-evenly;
   align-items: center;
+  width: 455px;
+  height: 362px;
+  text-align: center;
+  padding: 10px;
   .cardImage {
-    max-height: 75%;
-    width: 55%;
+    height: 260px;
+    width: 260px;
     border-radius: 1rem;
   }
   .innerCard {
@@ -34,5 +38,23 @@
   &:hover {
     cursor: pointer;
     transform: scale(1.05);
+  }
+}
+@media only screen and (max-width: 700px) {
+  .card {
+    width: 300px;
+    height: 300px;
+    .cardImage {
+      height: 200px;
+      width: 200px;
+    }
+    .innerCard {
+      p {
+        font-size: 15px;
+      }
+    }
+    .cardButton {
+      padding: 12px;
+    }
   }
 }
