+++ components/multiCards.module.scss
@@ -13,7 +13,7 @@
   }
 }
 
-@media only screen and (max-width: 700px) {
+@media only screen and (max-width: 800px) {
   .cardGrid {
     grid-template-columns: 90%;
   }
